chore: Update install, uninstall, and update scripts
Updated `install_pieye.sh` to include the option of passing a custom repository URL as the first argument. Additionally, added error checks to ensure that Git and Pip are installed on the system. The script now checks for the existence of the repository directory before cloning and installing the Pieye camera software. It enables the watchdog and sets up auto-login on the command line. Moreover, it copies `display_info.sh` to the `/home/pi` folder and adds an entry to execute `display_info.sh` on login.

In the `uninstall_pieye.sh` script, removed `display_info.sh` from the `/home/pi` folder and eliminated the corresponding entry from the `.profile` file.

In the `update_pieye.sh` script, updated the logic to check for local changes before updating the Pieye camera software, preventing updates when uncommitted changes exist. Error handling ensures that the Pieye service starts correctly after the update.

+++ scripts/update_pieye.sh
@@ -1,10 +1,25 @@
 #!/bin/bash
+set -e
+
+REPO_DIR="/home/pi/Pi-Eye"
+
+# Ensure directory exists
+if [ ! -d "$REPO_DIR" ]; then
+    echo "Directory $REPO_DIR does not exist. Exiting."
+    exit 1
+fi
 
 # Stop the service
 sudo systemctl stop pieye.service
 
 # Go to the repository
-cd /home/pi/Pi-Eye
+cd "$REPO_DIR"
+
+# Check for local changes
+if [ -n "$(git status -s)" ]; then
+    echo "Local changes found. Exiting."
+    exit 1
+fi
 
 # Get current version of package
 OLD_VERSION=$(python -c "import pieye; print(pieye.__version__)")
@@ -12,7 +27,7 @@ OLD_VERSION=$(python -c "import pieye; print(pieye.__version__)")
 # Update the repository
 git pull
 
-# Reinstall the Python package
+# Reinstall the Python package and its dependencies
 pip install . --upgrade
 
 # replace the .service file in systemd directory
@@ -23,6 +38,7 @@ sudo systemctl daemon-reload
 
 # Start the service
 sudo systemctl start pieye.service
+sudo systemctl is-active pieye.service || { echo "Failed to start pieye.service. Exiting."; exit 1; }
 
 # Get new version of package
 NEW_VERSION=$(python -c "import pieye; print(pieye.__version__)")
